requirements.txt
requests
python-dotenv
# Ajoutez d'autres dépendances ici si nécessaire
